Switch to Confluent Kafka as there is more documentation
